added file_result route and controller action to test executions controller. added file_result view which uses execution_results_for_file partial. move a lot of helper functions from test_executions_helper to test_execution model. TODO: move functions in test_execution model to a module, add controller tests for new controller action, and add integration tests for viewing uploaded xml on separate page. moved collected_errors() function and helper functions functions for this to an ErrorCollector module
