+++ app/controllers/test_executions_controller.rb
@@ -1,7 +1,9 @@
 class TestExecutionsController < ApplicationController
   include API::Controller
+  include TestExecutionsHelper
+  include Cypress::ErrorCollector
 
-  before_action :set_test_execution, only: [:show, :destroy]
+  before_action :set_test_execution, only: [:show, :destroy, :file_result]
   before_action :set_task, only: [:index, :new, :create]
   before_action :set_task_from_test_execution, only: [:show]
   before_action :set_product_test_from_task, only: [:show, :new]
@@ -48,6 +50,15 @@ class TestExecutionsController < ApplicationController
     render :nothing => true, :status => :no_content
   end
 
+  def file_result
+    @task = @test_execution.task
+    @product_test = @task.product_test
+    authorize! :read, @product_test.product.vendor
+    add_breadcrumbs
+    add_breadcrumb "File Results: #{route_file_name(params[:file_name])}"
+    file_name_and_error_result_from_execution(@test_execution)
+  end
+
   private
 
   def rescue_create
@@ -69,13 +80,28 @@ class TestExecutionsController < ApplicationController
     @product_test = @task.product_test
   end
 
+  def file_name_and_error_result_from_execution(execution)
+    errs = collected_errors(execution)
+    files = errs.files.select { |file_name, _| route_file_name(file_name) == params[:file_name] }
+    @file_name = files.first[0]
+    @error_result = files.first[1]
+  end
+
   def add_breadcrumbs
     add_breadcrumb 'Dashboard', :vendors_path
     add_breadcrumb 'Vendor: ' + @product_test.product.vendor.name, vendor_path(@product_test.product.vendor)
     add_breadcrumb 'Product: ' + @product_test.product.name, vendor_product_path(@product_test.product.vendor, @product_test.product)
     add_breadcrumb 'Checklist Dashboard: ' + @product_test.name, product_checklist_test_path(@product_test.product,
                                                                                              @product_test) if @product_test._type == 'ChecklistTest'
-    add_breadcrumb 'Test: ' + @product_test.name, product_product_test_path(@product_test.product, @product_test)
+    add_test_execution_breadcrumb
+  end
+
+  def add_test_execution_breadcrumb
+    if @task.most_recent_execution
+      add_breadcrumb 'Test: ' + @product_test.name, task_test_execution_path(task_id: @task.id, id: @task.most_recent_execution.id)
+    else
+      add_breadcrumb 'Test: ' + @product_test.name, ''
+    end
   end
 
   def results_params
@@ -85,5 +111,9 @@ class TestExecutionsController < ApplicationController
     elsif params[:test_execution]
       params[:test_execution][:results]
     end
+  end
+
+  def file_params
+    params.require(:file_name)
   end
 end
